pyproject.toml
[project]
name = "mf-register-bot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Kento Fukuda", email = "kento.fukuda@firstloop-tech.com" }
]
dependencies = [
    "ruff>=0.6.8",
    "pyppeteer>=2.0.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mf_register_bot"]
